34.4.B.0.252
Open source code for Xperia™ X Compact (SO-02J)

+++ drivers/soc/qcom/mpm-of.c
@@ -10,6 +10,11 @@
  * GNU General Public License for more details.
  *
  */
+/*
+ * NOTE: This file has been modified by Sony Mobile Communications Inc.
+ * Modifications are Copyright (c) 2017 Sony Mobile Communications Inc,
+ * and licensed under the license of the file.
+ */
 
 #include <linux/module.h>
 #include <linux/kernel.h>
@@ -535,7 +540,7 @@ bool msm_mpm_irqs_detectable(bool from_idle)
 			from_idle);
 }
 
-void msm_mpm_enter_sleep(uint32_t sclk_count, bool from_idle,
+void msm_mpm_enter_sleep(uint64_t sclk_count, bool from_idle,
 		const struct cpumask *cpumask)
 {
 	cycle_t wakeup = (u64)sclk_count * ARCH_TIMER_HZ;
